worker/requirements.txt
kafka-python
langchain-groq